feat: framework-independent `MakeswiftClient` implementation
