+++ .changeset/stupid-ducks-appear.md
@@ -1,0 +1,5 @@
+---
+'@makeswift/runtime': patch
+---
+
+feat: framework-independent `MakeswiftClient` implementation
