+++ packages/runtime/src/client/tests/client.error-handling.test.ts
@@ -1,5 +1,5 @@
 import { HttpResponse } from 'msw'
-import { failedResponseBody } from '../client'
+import { failedResponseBody } from '../../client'
 
 describe('failedResponseBody', () => {
   test('returns JSON for JSON responses', async () => {
