+++ packages/runtime/src/next/components/MakeswiftComponent.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { memo, Suspense, useMemo } from 'react'
-import { componentDocumentToRootEmbeddedDocument, MakeswiftComponentSnapshot } from '../client'
+import { componentDocumentToRootEmbeddedDocument, MakeswiftComponentSnapshot } from '../../client'
 import { DocumentRoot } from '../../runtimes/react/components/DocumentRoot'
 import { useCacheData } from '../../runtimes/react/hooks/use-cache-data'
 import { useRegisterDocument } from '../../runtimes/react/hooks/use-register-document'
