+++ packages/runtime/src/next/components/tests/makeswift-page-metadata-rendering.test.tsx
@@ -5,7 +5,7 @@ import { act } from 'react-dom/test-utils'
 import { render } from '@testing-library/react'
 import { ReactRuntime } from '../../../react'
 import * as Testing from '../../../runtimes/react/testing'
-import { type MakeswiftPageSnapshot, type MakeswiftPageDocument } from '../../client'
+import { type MakeswiftPageSnapshot, type MakeswiftPageDocument } from '../../../client'
 import { type CacheData } from '../../../api/react'
 import { Page as MakeswiftPage } from '../page'
 import { ComponentPropsWithoutRef } from 'react'
