+++ packages/runtime/src/next/index.tsx
@@ -2,7 +2,7 @@ export type { PageProps } from './components/page'
 export { Page } from './components/page'
 export { MakeswiftComponent } from './components/MakeswiftComponent'
 export { Slot } from './components/Slot'
-export type { MakeswiftPage, MakeswiftPageDocument, MakeswiftPageSnapshot, Sitemap } from './client'
+export type { MakeswiftPage, MakeswiftPageDocument, MakeswiftPageSnapshot, Sitemap } from '../client'
 export { Makeswift } from './client'
 export type { MakeswiftPreviewData } from './preview-mode'
 export { RootStyleRegistry } from './root-style-registry'
